+++ app/src/main/java/com/poupa/vinylmusicplayer/appwidgets/AppWidgetSmall.java
@@ -5,7 +5,7 @@ import android.widget.RemoteViews;
 
 import com.poupa.vinylmusicplayer.R;
 import com.poupa.vinylmusicplayer.appwidgets.base.BaseAppWidget;
-import com.poupa.vinylmusicplayer.discog.MultiValuesTagUtil;
+import com.poupa.vinylmusicplayer.discog.tags.MultiValuesTagUtil;
 import com.poupa.vinylmusicplayer.model.Song;
 import com.poupa.vinylmusicplayer.service.MusicService;
