+++ app/src/main/java/com/poupa/vinylmusicplayer/discog/DB.java
@@ -8,6 +8,7 @@ import android.database.sqlite.SQLiteOpenHelper;
 import androidx.annotation.NonNull;
 
 import com.poupa.vinylmusicplayer.App;
+import com.poupa.vinylmusicplayer.discog.tags.MultiValuesTagUtil;
 import com.poupa.vinylmusicplayer.model.Song;
 
 import java.util.ArrayList;
