+++ app/src/main/java/com/poupa/vinylmusicplayer/helper/M3UWriter.java
@@ -2,7 +2,7 @@ package com.poupa.vinylmusicplayer.helper;
 
 import android.content.Context;
 
-import com.poupa.vinylmusicplayer.discog.MultiValuesTagUtil;
+import com.poupa.vinylmusicplayer.discog.tags.MultiValuesTagUtil;
 import com.poupa.vinylmusicplayer.model.Playlist;
 import com.poupa.vinylmusicplayer.model.Song;
